Rename InitialAvatar file to UserAvatar
