+++ packages/front-end/components/Avatar/UserAvatar.tsx
